requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
flatbuffers==2.0
idna==3.3
invisible-watermark==0.1.5
numpy==1.22.2
onnx==1.10.2
onnxruntime==1.10.0
opencv-python==4.5.5.62
Pillow
protobuf==3.19.4
pyTelegramBotAPI==4.4.0
PyWavelets==1.2.0
requests==2.27.1
six==1.16.0
torch==1.10.2
typing-extensions==4.0.1
urllib3==1.26.8